+++ src/Firebase/Messaging/MessageData.php
@@ -27,6 +27,13 @@ final class MessageData implements \JsonSerializable
                 throw new InvalidArgumentException('Message data must be a one-dimensional array of string(able) keys and values.');
             }
 
+            if (self::isBinary((string) $value)) {
+                throw new InvalidArgumentException(
+                    "The message data field '{$key}' seems to contain binary data. As this can lead to broken messages, "
+                    .'please convert it to a string representation first, e.g. with bin2hex() or base64encode().'
+                );
+            }
+
             $messageData->data[(string) $key] = (string) $value;
         }
 
@@ -47,5 +54,10 @@ final class MessageData implements \JsonSerializable
     private static function isStringable($value): bool
     {
         return \is_null($value) || \is_scalar($value) || (\is_object($value) && \method_exists($value, '__toString'));
+    }
+
+    private static function isBinary(string $value): bool
+    {
+        return \mb_detect_encoding($value) === false;
     }
 }