+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG
 
 ## [Unreleased]
+### Changed
+* Message data added to a with `CloudMessage::withData()` now rejects binary data to avoid broken
+  messages being sent to the Firebase API.
+  ([#441](https://github.com/kreait/firebase-php/issues/441))
 
 ## [5.5.0] - 2020-06-19
 ### Added